requirements.txt
python-telegram-bot==21.4
httpx==0.27.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
